Move the models into the lib
