+++ catgrad/Cargo.toml
@@ -18,17 +18,14 @@ test-log = "0.2.17"
 memmap2 = "0.9.5"
 rayon = "1.10.0"
 num_cpus = "1.17.0"
+serde = { version = "1.0.219", features = ["derive"] }
+serde_json = "1.0.140"
 
 [dev-dependencies]
 clap = { version = "4.5.37", features = ["derive"] }
 env_logger = "0.11.8"
 hf-hub = { version = "0.4.2", default-features = false, features = ["ureq"] }
 safetensors = "0.5.3"
-serde = { version = "1.0.219", features = ["derive"] }
-serde_json = "1.0.140"
 tokenizers = { version = "0.21.1", features = ["http"] }
 open-hypergraphs-dot = "0.2.1"
 graphviz-rust = "0.9.5"
-
-[profile.release]
-debug = true
