+++ catgrad/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod backend;
 pub mod core;
+pub mod llm;
