+++ catgrad/src/llm/models/llama.rs
@@ -1,9 +1,9 @@
-// OLMo-2 model description
+// Llama-3 model description
 
-use super::{Cache, Config, ModelBuilder};
-use catgrad::backend::cpu::eval::Builder;
-use catgrad::core::nn::layers::*;
-use catgrad::core::{Dtype, NdArrayType, Shape, Var};
+use super::utils::{Cache, Config, ModelBuilder};
+use crate::backend::cpu::eval::Builder;
+use crate::core::nn::layers::*;
+use crate::core::{Dtype, NdArrayType, Shape, Var};
 
 pub struct Model;
 
@@ -18,6 +18,7 @@ impl ModelBuilder for Model {
     ) -> Var {
         let tokens = x.label.shape.0[1];
         let emb = Model::embeddings(builder, config, x);
+
         let mut result = emb;
 
         for i in 0..config.num_hidden_layers {
@@ -39,13 +40,17 @@ impl ModelBuilder for Model {
             result = narrow(builder, 1, tokens - 1, 1, result);
         }
 
-        linear_no_bias(
-            builder,
-            config.hidden_size,
-            config.vocab_size,
-            "lm_head",
-            result,
-        )
+        // Add lm_head if weight tying is used
+        if config.tie_word_embeddings {
+            result = linear_no_bias(
+                builder,
+                config.hidden_size,
+                config.vocab_size,
+                "model.embed_tokens",
+                result,
+            );
+        }
+        result
     }
 }
 
@@ -71,6 +76,7 @@ impl Model {
         let dim = config.hidden_size;
         let num_heads = config.num_attention_heads;
         let num_kv_heads = config.num_key_value_heads;
+        let rep = num_heads / num_kv_heads;
         let head_dim = config.hidden_size / num_heads;
         let b = x.label.shape.0[0];
         let s = x.label.shape.0[1];
@@ -79,20 +85,11 @@ impl Model {
         let k = linear_no_bias(
             builder,
             dim,
-            dim * num_kv_heads / num_heads,
+            dim / rep,
             &format!("{name}.k_proj"),
             x.clone(),
         );
-        let v = linear_no_bias(
-            builder,
-            dim,
-            dim * num_kv_heads / num_heads,
-            &format!("{name}.v_proj"),
-            x,
-        );
-
-        let q = rmsnorm(builder, config.rms_norm_eps, &format!("{name}.q_norm"), q);
-        let k = rmsnorm(builder, config.rms_norm_eps, &format!("{name}.k_norm"), k);
+        let v = linear_no_bias(builder, dim, dim / rep, &format!("{name}.v_proj"), x);
 
         let q = reshape(builder, Shape(vec![b, s, num_heads, head_dim]), q);
         let k = reshape(builder, Shape(vec![b, s, num_kv_heads, head_dim]), k);
@@ -112,6 +109,9 @@ impl Model {
             }
             cache.kv_cache[layer_id] = Some((k.clone(), v.clone()));
         };
+
+        let k = repeat_kv(builder, rep, k);
+        let v = repeat_kv(builder, rep, v);
 
         let tk = transpose(builder, 2, 3, k);
         let attn = mat_mul(builder, q, tk);
@@ -166,6 +166,12 @@ impl Model {
         x: Var,
     ) -> Var {
         let res = x.clone();
+        let x = rmsnorm(
+            builder,
+            config.rms_norm_eps,
+            &format!("{name}.input_layernorm"),
+            x,
+        );
         let x = Model::attention(
             builder,
             layer_id,
@@ -175,22 +181,15 @@ impl Model {
             &format!("{name}.self_attn"),
             x,
         );
+        let x = res + x;
+        let res = x.clone();
         let x = rmsnorm(
             builder,
             config.rms_norm_eps,
             &format!("{name}.post_attention_layernorm"),
             x,
         );
-        let x = res + x;
-
-        let res = x.clone();
         let x = Model::mlp(builder, config, &format!("{name}.mlp"), x);
-        let x = rmsnorm(
-            builder,
-            config.rms_norm_eps,
-            &format!("{name}.post_feedforward_layernorm"),
-            x,
-        );
         x + res
     }
 }
